provider/librato: Fixes for various integer type casting bugs
Fixes #8968

+++ builtin/providers/librato/resource_librato_alert.go
@@ -182,8 +182,8 @@ func resourceLibratoAlertCreate(d *schema.ResourceData, meta interface{}) error 
 			if v, ok := conditionData["detect_reset"].(bool); ok {
 				condition.DetectReset = librato.Bool(v)
 			}
-			if v, ok := conditionData["duration"].(uint); ok {
-				condition.Duration = librato.Uint(v)
+			if v, ok := conditionData["duration"].(int); ok {
+				condition.Duration = librato.Uint(uint(v))
 			}
 			if v, ok := conditionData["summary_function"].(string); ok && v != "" {
 				condition.SummaryFunction = librato.String(v)
@@ -343,7 +343,7 @@ func resourceLibratoAlertUpdate(d *schema.ResourceData, meta interface{}) error 
 		alert.Active = librato.Bool(d.Get("active").(bool))
 	}
 	if d.HasChange("rearm_seconds") {
-		alert.RearmSeconds = librato.Uint(d.Get("rearm_seconds").(uint))
+		alert.RearmSeconds = librato.Uint(uint(d.Get("rearm_seconds").(int)))
 	}
 	if d.HasChange("services") {
 		vs := d.Get("services").(*schema.Set)
@@ -374,8 +374,8 @@ func resourceLibratoAlertUpdate(d *schema.ResourceData, meta interface{}) error 
 		if v, ok := conditionData["detect_reset"].(bool); ok {
 			condition.DetectReset = librato.Bool(v)
 		}
-		if v, ok := conditionData["duration"].(uint); ok {
-			condition.Duration = librato.Uint(v)
+		if v, ok := conditionData["duration"].(int); ok {
+			condition.Duration = librato.Uint(uint(v))
 		}
 		if v, ok := conditionData["summary_function"].(string); ok && v != "" {
 			condition.SummaryFunction = librato.String(v)
